pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jobbot"
version = "0.1.0"
description = "POC job-search automation that collects jobs and sends Telegram digests."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "PyYAML>=6.0",
  "requests>=2.31.0",
  "beautifulsoup4>=4.12.0",
  "python-dotenv>=1.0.0",
  "playwright>=1.44.0",
]

[project.scripts]
jobbot = "jobbot.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

